Renaming from SP_aircraft to SPaircraft for consistency.
