+++ SPaircraft.py
