+++ TESTS
@@ -1,1 +1,1 @@
-SP_aircraft.py
+SPaircraft.py
